+++ packages/core/ui/styling/css-shadow.ts
@@ -1,6 +1,6 @@
-import { CoreTypes } from '../../core-types';
 import { Color } from '../../color';
-import { Length } from './style-properties';
+import { CoreTypes } from '../../core-types';
+import { parseCSSShorthand } from './css-utils';
 
 export interface ShadowCSSValues {
 	inset: boolean;
@@ -11,55 +11,6 @@ export interface ShadowCSSValues {
 	color: Color;
 }
 
-/**
- * Matches whitespace except if the whitespace is contained in parenthesis - ex. rgb(a), hsl color.
- */
-const PARTS_RE = /\s(?![^(]*\))/;
-
-/**
- * Matches a Length value with or without a unit
- */
-const LENGTH_RE = /^-?[0-9]+[a-zA-Z%]*?$/;
-
-/**
- * Checks if the value is a Length or 0
- */
-const isLength = (v) => v === '0' || LENGTH_RE.test(v);
-
-export function parseCSSShorthand(value: string): {
-	values: Array<CoreTypes.LengthType>;
-	color: string;
-	inset: boolean;
-} {
-	const parts = value.trim().split(PARTS_RE);
-	const inset = parts.includes('inset');
-	const first = parts[0];
-	const last = parts[parts.length - 1];
-
-	let color = 'black';
-	if (first !== 'none') {
-		if (first && !isLength(first) && first !== 'inset') {
-			color = first;
-		} else if (last && !isLength(last)) {
-			color = last;
-		}
-	}
-	const values = parts
-		.filter((n) => n !== 'inset')
-		.filter((n) => n !== color)
-		.map((val) => {
-			try {
-				return Length.parse(val);
-			} catch (err) {
-				return CoreTypes.zeroLength;
-			}
-		});
-	return {
-		inset,
-		color,
-		values,
-	};
-}
 /**
  * Parse a string into ShadowCSSValues
  * Supports any valid css box/text shadow combination.
@@ -78,6 +29,6 @@ export function parseCSSShadow(value: string): ShadowCSSValues {
 		offsetY: offsetY,
 		blurRadius: blurRadius,
 		spreadRadius: spreadRadius,
-		color: new Color(data.color),
+		color: data.color ? new Color(data.color) : undefined,
 	};
 }
